+++ riscv/insns/vsmul_vx.h
@@ -1,8 +1,8 @@
 // vsmul
 VRM xrm = P.VU.get_vround_mode();
-uint128_t int_max = (1ul << (P.VU.vsew - 1)) - 1;
+uint128_t int_max = (uint64_t(1) << (P.VU.vsew - 1)) - 1;
 uint128_t int_min = - (1 << (P.VU.vsew - 1));
-uint128_t sign_mask = ((1ul << (P.VU.vsew - 1)));
+uint128_t sign_mask = uint64_t(1) << (P.VU.vsew - 1);
 
 VI_VX_ULOOP
 ({
